Correctly import Debug, factor out mapper
